Move `Message` & `IncomingIndex` from `peerset` to `protocol_controller`
